pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Nazar Chaban <134311145+Vivern0@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.9.5"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
aiohttp-jinja2 = "^1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
